Rename DistanceMetric sub-classes
